Add files via upload
Updated functionality; added the ability to double-click employee names on the schedule, which opens up a new window and the user can indicate that the employee is sick, which is then added to the Employee object's sick time. Added the ability to denote the date of the schedule's first day, which is then reflected in the Schedule tab's column labels.